requirements.txt
flask
pyngrok 
twilio 
gspread 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
dotenv
requests
streamlit
